LayoutTests/inspector/codemirror/prettyprinting-css.html

<!DOCTYPE html>
<html>
<head>
<script src="../../http/tests/inspector/resources/inspector-test.js"></script>
<script src="resources/prettyprinting/utilities.js"></script>
<script>
function test()
{
    let suite = InspectorTest.createAsyncSuite("CodeMirror.PrettyPrinting.CSS");

    addPrettyPrintingTests(suite, "text/css", [
        "resources/prettyprinting/css-tests/basic.css",
        "resources/prettyprinting/css-tests/calc.css",
        "resources/prettyprinting/css-tests/gradient.css",
        "resources/prettyprinting/css-tests/keyframes.css",
        "resources/prettyprinting/css-tests/media-query.css",
        "resources/prettyprinting/css-tests/selectors.css",
        "resources/prettyprinting/css-tests/wrapping.css",
    ]);

    suite.runTestCasesAndFinish();
}
</script>
</head>
<body onload="runTest()">
<p>Test CSS pretty printing with CodeMirror.</p>
</body>
</html>
